refactor(runner): flatten packages runner, config, report
